# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "1.0.0",
    "description": "เอกสารระบบ API สำหรับจัดการ Todo List"
  },
  "components": {
    "securitySchemes": {
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "ยิง GET /api/csrf-token แล้วนำค่า csrfToken มาใส่ที่นี่"
      }
    }
  },
  "security": [ { "csrfToken": [] } ],
  "servers": [ { "url": "http://localhost:3000" } ],
  "paths": {
...
    "/api/todos": {
      "post": {
        "summary": "สร้าง Todo ใหม่",
        "tags": ["Todos"],
        "security": [{ "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "เขียน Document" },
                  "description": { "type": "string", "example": "ทำเอกสาร API" },
                  "status": { "type": "string", "example": "PENDING" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "สร้างสำเร็จ" } }
      },
      "get": {
        "summary": "ดึงรายการ Todo ทั้งหมด",
        "tags": ["Todos"],
        "security": [{ "csrfToken": [] }],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer" }, "description": "หน้าที่ต้องการดู" },
          { "in": "query", "name": "limit", "schema": { "type": "integer" }, "description": "จำนวนรายการต่อหน้า" },
          { "in": "query", "name": "search", "schema": { "type": "string" }, "description": "ค้นหาจากชื่อ Title" },
          { "in": "query", "name": "status", "schema": { "type": "string" }, "description": "สถานะ (PENDING, IN_PROGRESS, DONE)" }
        ],
        "responses": { "200": { "description": "ดึงข้อมูลสำเร็จ" } }
      }
    },
...
  }
}